+++ modules/engine/lib/engine/insert.js
@@ -20,7 +20,7 @@ var jsonfill = require('./jsonfill.js'),
     _ = require('underscore'),
     assert = require('assert');
 
-exports.exec = function(opts, statement, cb, parentEvent) {
+exports.exec = function(opts, statement, parentEvent, cb) {
 
     assert.ok(opts.tables, 'Argument tables can not be undefined');
     assert.ok(statement, 'Argument statement can not be undefined');